add a RunbotCaseMinimalSetup
